Add ASAN hooks to poison bumpalloc heap, fix RDI use on Win
When running under -fsanitize=address, hook into the poison/unpoison
runtime routines to get bounds checking (and leak checking on Linux).

Using ASAN on Windows caused a mostly-unrelated crash caused by
incorrect use of RDI during codegen -- I was treating it as volatile,
and a utility register, but on Windows it's non-volatile which the ASAN
runtime cleanup after main() happened to rely on due to lucky
optimizer-chance. (It's kind of unfortunate for 'rep stosb' usage too,
which is used to initialize locals, perhaps should revisit that.) RCX is
now used on Windows instead for a default utility volatile register.

+++ unicode.c
@@ -68,7 +68,7 @@ uint32_t decode_utf8(char** new_pos, char* p) {
 }
 
 static bool in_range(uint32_t* range, uint32_t c) {
-  for (int i = 0; range[i] != -1; i += 2)
+  for (uint32_t i = 0; range[i] != (uint32_t)-1; i += 2)
     if (range[i] <= c && c <= range[i + 1])
       return true;
   return false;
